Remove repetitive Task classes from subtasks, inherit from _base_task.py instead.
